algumas mudanças3
